[ADD] stock_maintenance: display lots
Commit [1] is using models and data from `stock` module but
`maintenance` does not depend on it.

[1] 1f2a5806b832529bd2ec53a878991e04800079bf

sentry-5964505411
