+++ addons/maintenance/views/maintenance_views.xml
@@ -370,12 +370,6 @@
                 <sheet>
                     <field name="company_id" invisible="1"/>
                     <div class="oe_button_box" name="button_box">
-                        <button type="object"
-                            name="action_open_matched_serial"
-                            icon="fa-bars" class="oe_stat_button"
-                            invisible="not match_serial">
-                            <field string="Serial Number" name="serial_no" widget="statinfo"/>
-                        </button>
                         <button name="%(hr_equipment_request_action_from_equipment)d"
                             type="action"
                             class="oe_stat_button"
